+++ spec/components/service_navigation_spec.rb
@@ -142,4 +142,34 @@ describe "Service navigation", type: :view do
                        collapse_navigation_on_mobile: true })
     assert_select(".govuk-service-navigation__toggle", text: "Menu")
   end
+
+  it "renders the hide_service_name_on_mobile option correctly" do
+    render_component({
+      navigation_items: [
+        {
+          text: "Navigation item 1",
+          href: "#",
+        },
+      ],
+      hide_service_name_on_mobile: true,
+      service_name: "My service home",
+      service_name_url: "/home",
+      toggle_text: "My service",
+    })
+    assert_select(".gem-c-service-navigation--hide-service-name-on-mobile")
+  end
+
+  it "renders the columns_layout option correctly" do
+    render_component({
+      navigation_items: [
+        {
+          text: "Navigation item 1",
+          href: "#",
+        },
+      ],
+      columns_layout: true,
+      service_name: "My service home",
+    })
+    assert_select(".gem-c-service-navigation--columns-layout")
+  end
 end